Seguridad

+++ src/team/team.controller.ts
@@ -1,10 +1,12 @@
-import { Body, Controller, Delete, Get, Inject, Param, Post, Put, UseInterceptors } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Inject, Param, Post, Put, Req, UnauthorizedException, UseGuards, UseInterceptors } from '@nestjs/common';
 import { ApiBearerAuth, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { TeamService } from './team.service';
 import { Cache, CACHE_MANAGER, CacheInterceptor } from '@nestjs/cache-manager';
 import { UpdateTeamDto } from './dto/update-team.dto';
 import { CreateTeamDto } from './dto/create-team.dto';
 import { Team } from './schema/team.schema';
+import { AuthGuard } from '@nestjs/passport';
+import { Role } from 'src/auth/schemas/user.schema';
 
 @Controller('teams')
 @ApiTags('Teams')
@@ -15,11 +17,19 @@ export class TeamController {
     ) {}
 
     @Post()
+    @UseGuards(AuthGuard('jwt'))
+    @ApiBearerAuth()
     @ApiOperation({ summary: 'Crear un nuevo equipo' }) 
     async create(
+        @Req() req,
         @Body() createTeamDto: CreateTeamDto
     ): Promise<Team> {
-        return this.teamService.create(createTeamDto);
+        const user = req.user;
+        if (!user || user.role == Role.USER)
+            throw new UnauthorizedException('You must be logged as administrator in to create a competition');
+        else {
+            return this.teamService.create(createTeamDto);
+        }
     }
 
     @Get()
@@ -48,19 +58,35 @@ export class TeamController {
     }
 
     @Put(':id')
+    @UseGuards(AuthGuard('jwt'))
+    @ApiBearerAuth()
     @ApiOperation({ summary: 'Actualizar un equipo' })
     async update(
+        @Req() req,
         @Param('id') id: string,
         @Body() updateTeamDto: UpdateTeamDto
     ): Promise<Team> {
-        return this.teamService.update(id, updateTeamDto);
+        const user = req.user;
+        if (!user || user.role == Role.USER)
+            throw new UnauthorizedException('You must be logged as administrator in to create a competition');
+        else {
+            return this.teamService.update(id, updateTeamDto);
+        }
     }
 
     @Delete(':id')
+    @UseGuards(AuthGuard('jwt'))
+    @ApiBearerAuth()
     @ApiOperation({ summary: 'Eliminar un equipo' })
     async remove(
+        @Req() req,
         @Param('id') id: string
     ): Promise<void> {
-        await this.teamService.remove(id);
+        const user = req.user;
+        if (!user || user.role == Role.USER)
+            throw new UnauthorizedException('You must be logged as administrator in to create a competition');
+        else {
+            await this.teamService.remove(id);
+        }
     }
 }
